Adapt choser to use the new shared component
